user update kilometer change int to double

+++ app/src/main/java/miage/parisnanterre/fr/runwithme/User.java
@@ -37,7 +37,7 @@ public class User {
         return km;
     }
 
-    public void setKm(int km) {
+    public void setKm(double km) {
         this.km = km;
     }
     public void updateKm(double u){
